CerebroAdmin/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
asyncpg==0.30.0
pydantic==2.10.5
pydantic-settings==2.7.1
python-dotenv==1.0.1
google-generativeai==0.8.3
httpx==0.28.1
